Move provable_height_tracker to a submodule
